requirements.txt
# -e .
aws-cdk.core==1.106.1
aws-cdk.aws-s3==1.106.1
aws-cdk.aws-iam==1.106.1